// src/interfaces/AnkiFields.ts
export interface AnkiFields {
    audioFile: string;
    audioUrl: string;
    dictionaryForm: string;
    word: string;
    reading: string;
    saveNotes: null;
    sentence: string;
    sentenceWithBlank: string;
    sourceUrl: string;
    pageTitle: string;
    definition: string;
    frequency: string;
    pitch: string;
}